requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.12.0
python-multipart>=0.0.5
pydantic>=1.8.2
scikit-learn>=0.24.2
nltk>=3.6.3
pandas>=1.3.3
requests>=2.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
transformers>=4.11.3
torch>=1.9.0
openai>=1.0.0
stripe>=5.0.0
pyjwt>=2.3.0